refactor(statistics): extract shared useMonthKeyFormatter hook
Three chart components were duplicating an identical memoized
Intl.DateTimeFormat helper that parses "YYYY-MM" keys. Extract to a
single hook so the format and locale handling live in one place.
